refactor(ONB-205): Refactor store management seperately.
